Not anymore a widget
It's a harder to mantain both a widget and an app, as I have to code everything twice, and in the widget it wouldn't be possible to use buttonListener so I have to drop it. Also this will reduce file size, not by much but installation will be faster and won't kill launcher every time you update it.

Signed-off-by: Miguel <micrusa@hotmail.es>